requirements.txt
blockfrost-python==0.5.3
click==8.1.3
requests==2.29.0
